% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvaluator.R
\name{dvaluator}
\alias{dvaluator}
\title{Dependent Variable Evaluator Function}
\usage{
dvaluator(depvar, order = F, panel = F)
}
\arguments{
\item{depvar}{What is the name of your dependent variable?}

\item{order}{Is your variable ordinal? Default is set to FALSE}

\item{panel}{Is your data in panel or cross sectional form? Default is set to FALSE}
}
\description{
This function's main use is to check the nature of the dependent variable, suggest a model type, and explain the model to the user.
}
\examples{

}
