% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{linear_regression}
\alias{linear_regression}
\title{Compute linear regression on data using formula and make it work with magrittr pipe operator.}
\usage{
linear_regression(data, formula)
}
\arguments{
\item{data}{A dataframe, e.g. mtcars.}

\item{formula}{A formula, e.g. mpg ~ wt.}
}
\value{
A modified dataframe containing the variables in the equation
plus the fitted value and the residuals.  The formula, model coefficients,
variable correlation matrix and overall R^2 and AdjR^2 are attached to
the dataframe as attributes.
}
\description{
Compute linear regression on data using formula and make it work with magrittr pipe operator.
}
\examples{
library(datasets)
library(dplyr)
library(regressR)
result <- linear_regression(mtcars, mpg ~ wt)
result2 <- mtcars \%>\% linear_regression(mpg ~ wt)
}
