% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.tester.R
\name{package.tester}
\alias{package.tester}
\title{Package Testing Function}
\usage{
package.tester(package)
}
\arguments{
\item{package}{Here you can inpute a vector of names of packages you wish to install/run on this computer in 1 go.}
}
\description{
This function tests whether certain packages are installed on this R instance for some of the other functions to run. It 
It can also serve as a way for the user to test for other packaeges
}
\examples{
package.tester()
}
\keyword{package}
