% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizer.R
\name{summarizer}
\alias{summarizer}
\title{A summary function}
\usage{
summarizer(df)
}
\arguments{
\item{df}{This is where a data frame goes}
}
\description{
This function allows you to express your love of cats.
}
\examples{
summarizer()
}
\keyword{basic}
