\name{winemag_data_130k_v2}
\alias{winemag_data_130k_v2}
\docType{data}
\title{
Wine Magazine wine ratings
}
\description{
Wine Magazine wine ratings
}
\usage{data("winemag_data_130k_v2")}
\format{
  A data frame with 129971 observations on the following 14 variables.
  \describe{
    \item{\code{X1}}{a numeric vector}
    \item{\code{country}}{a character vector}
    \item{\code{description}}{a character vector}
    \item{\code{designation}}{a character vector}
    \item{\code{points}}{a numeric vector}
    \item{\code{price}}{a numeric vector}
    \item{\code{province}}{a character vector}
    \item{\code{region_1}}{a character vector}
    \item{\code{region_2}}{a character vector}
    \item{\code{taster_name}}{a character vector}
    \item{\code{taster_twitter_handle}}{a character vector}
    \item{\code{title}}{a character vector}
    \item{\code{variety}}{a character vector}
    \item{\code{winery}}{a character vector}
  }
}
\details{
Wine Magazine wine ratings
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(winemag_data_130k_v2)
## maybe str(winemag_data_130k_v2) ; plot(winemag_data_130k_v2) ...
}
\keyword{datasets}
