% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizer.R
\name{summarizer}
\alias{summarizer}
\title{A summary function}
\usage{
summarizer(df)
}
\arguments{
\item{df}{This is where a data frame goes. Not a lot of science here.}
}
\description{
This function gives someone a sky-in-the-eye view of the data as it looks into the structure of the data frame and gives reccomendations.
}
\examples{
summarizer()
}
\keyword{statistics}
\keyword{summary}
